+++ src/interaction.ts
@@ -1,3 +1,5 @@
+import { hasParentNode } from "./dom";
+
 export interface InteractionInfo {
   c?: number; // Click time
   ci?: string; // Click attribution identifier
@@ -6,4 +8,62 @@ export interface InteractionInfo {
   k?: number; // Key press time
   ki?: string; // Key press attribution identifier
   s?: number; // Scroll time
+}
+
+type ButtonOrLinkElement = HTMLButtonElement | HTMLLinkElement;
+
+/**
+ * Get the interaction attribution name for an element
+ *
+ * @param {HTMLElement} el
+ * @returns string
+ */
+export function interactionAttributionForElement(el: HTMLElement): string {
+  // Our first preference is to use the data-sctrack attribute from anywhere in the tree
+  const trackId = getClosestScTrackAttribute(el);
+
+  if (trackId) {
+    return trackId;
+  }
+
+  // The second preference is to use the element's ID
+  if (el.id) {
+    return el.id;
+  }
+
+  // The third preference is to use the text content of a button or link
+  const isSubmitInput = el.tagName === "INPUT" && (el as HTMLInputElement).type === "submit";
+  const isButton = el.tagName === "BUTTON";
+  const isLink = el.tagName === "A";
+
+  if (isSubmitInput && (el as HTMLInputElement).value) {
+    return (el as HTMLInputElement).value;
+  }
+
+  if ((isButton || isLink) && (el as ButtonOrLinkElement).innerText) {
+    return (el as ButtonOrLinkElement).innerText;
+  }
+
+  if (hasParentNode(el)) {
+    return interactionAttributionForElement(el.parentNode as HTMLElement);
+  }
+
+  // No suitable attribute was found
+  return "";
+}
+
+function getClosestScTrackAttribute(el: HTMLElement): string | null {
+  if (el.hasAttribute("data-sctrack")) {
+    const trackId = el.getAttribute("data-sctrack")?.trim();
+
+    if (trackId) {
+      return trackId;
+    }
+  }
+
+  if (hasParentNode(el)) {
+    return getClosestScTrackAttribute(el.parentNode as HTMLElement);
+  }
+
+  return null;
 }